[SERVICES] Switch Extractor to Services
