+++ sv_dlp/services/apple/proto/ResourceManifest.proto
